feat: add OutputProcessor for file handling
