+++ src/agent/output/IOutputHandler.ts
@@ -1,6 +1,7 @@
 // Local imports - types
 import { NamedOutputFile, OutputFileInfo } from './types';
 import { XmlOutputManager } from './XmlOutputManager';
+import { OutputProcessor } from './OutputProcessor';
 
 /** Interface describing OutputHandler behavior used by agents. */
 export interface IOutputHandler {
@@ -10,14 +11,11 @@ export interface IOutputHandler {
   /** Mapping of source to processed output files by round. */
   outputMappings: { [key: number]: NamedOutputFile[] };
 
+  /** Processor for splitting and formatting outputs. */
+  outputProcessor: OutputProcessor;
+
   /** XML manager for parsing and splitting outputs. */
   xmlManager: XmlOutputManager;
-
-  /** Indent a single LaTeX file for readability. */
-  indentLatexFile(filePath: string): Promise<void>;
-
-  /** Indent multiple LaTeX files for readability. */
-  indentLatexFiles(filePaths: string[]): Promise<void>;
 
   /** Run latexdiff comparisons for the current round. */
   handleLatexdiffofOutput(currRound: number, groupId?: string): Promise<void>;
