+++ src/agent/output/index.ts
@@ -1,4 +1,5 @@
 export * from './OutputHandler';
 export * from './IOutputHandler';
+export * from './OutputProcessor';
 export { NamedOutputFile } from './types';
 export { getOutputFileName } from '@agent/utils/outputFileUtils';
